+++ server/test/tuist/billing_test.exs
@@ -15,18 +15,18 @@ defmodule Tuist.BillingTest do
   setup do
     stub(Environment, :stripe_prices, fn ->
       %{
-        air: %{
-          usage: ["air.usage"],
-          flat_monthly: ["air.flat.monthly"]
+        "air" => %{
+          "usage" => ["air.usage"],
+          "flat_monthly" => ["air.flat.monthly"]
         },
-        pro: %{
-          usage: ["pro.usage"],
-          flat_monthly: ["pro.flat.monthly"],
-          flat_yearly: ["pro.flat.yearly"]
+        "pro" => %{
+          "usage" => ["pro.usage"],
+          "flat_monthly" => ["pro.flat.monthly"],
+          "flat_yearly" => ["pro.flat.yearly"]
         },
-        enterprise: %{
-          flat_monthly: ["enterprise.flat.monthly"],
-          flat_yearly: ["enterprise.flat.yearly"]
+        "enterprise" => %{
+          "flat_monthly" => ["enterprise.flat.monthly"],
+          "flat_yearly" => ["enterprise.flat.yearly"]
         }
       }
     end)
@@ -323,7 +323,9 @@ defmodule Tuist.BillingTest do
         items: %{data: [%{price: %{id: "pro.usage"}}, %{price: %{id: "pro.flat.monthly"}}]}
       })
 
-      stub(Stripe.Subscription, :cancel, fn "sub_some-id" -> {:ok, %Stripe.Subscription{status: "canceled"}} end)
+      stub(Stripe.Subscription, :cancel, fn "sub_some-id" ->
+        {:ok, %Stripe.Subscription{status: "canceled"}}
+      end)
 
       # When
       Billing.on_subscription_change(%{
@@ -454,7 +456,10 @@ defmodule Tuist.BillingTest do
                                                method: :post,
                                                endpoint: "/v1/billing/meter_events",
                                                params: %{
-                                                 payload: %{value: 10, stripe_customer_id: ^customer_id},
+                                                 payload: %{
+                                                   value: 10,
+                                                   stripe_customer_id: ^customer_id
+                                                 },
                                                  event_name: "remote_cache_hit"
                                                }
                                              } ->